Move GameManagement contents to Model
Yesod essentially assumes that Model is where all of the data modeling
goes. Move logic there rather than swim against the tide.
